crt: Prefer ucrt's math functions over libmingwex, for arch independent functions
The ucrt math functions are C99 compliant, and should in general
be usable.

Move such symbols from src_libmingwex to src_msvcrt_common. By not
having them in libmingwex, the corresponding imported functions from
libucrtbase.a/libucrt.a will be preferred instead.

First do this for source files that are handled identically for all
architectures.

Signed-off-by: Martin Storsjö <martin@martin.st>