headers: Fix _stprintf and _vstprintf macros in _UNICODE mode
Per MS documentation, macros _stprintf and _vstprintf always expands to
s(w)printf function variant which does not take buffer size argument.

https://learn.microsoft.com/en-us/cpp/c-runtime-library/reference/sprintf-sprintf-l-swprintf-swprintf-l-swprintf-l

So _stprintf() expands either to C89 sprintf() function or to MS-specific
_swprintf() function. C95 swprintf() takes buffer size argument and so, it
is not compatible for _stprintf() and cannot be used.

So replace swprintf by _swprintf in _UNICODE mode of _stprintf expansion.

Signed-off-by: Martin Storsjö <martin@martin.st>