Add `-mno-outline-atomics` to `CFLAGS`` when building `mingw-w64-crt` and `mingw-w64-libraries/winpthreads` for `aarch64-*`
As a follow up to the first patch "[PATCH] Add check if `libgcc` is available and link `winpthreads` against it instead of `fakelib`" I am sending another possibility how to overcome the undefined references of to outline atomics LSE intrinsic functions for `aarch64-w64-mingw32` GCC build that adds `-mno-outline-atomics` flag to build of `mingw-w64-crt` and `mingw-w64-libraries/winpthreads`. Note that the flag is required for both libraries.

Signed-off-by: Martin Storsjö <martin@martin.st>