github: Run test-ffmpeg on arm too
Building this on arm requires passing the --host option; on x86
the default "uname -m" works for both i686 and x86_64 (ffmpeg's
configure maps this to the architecture "x86", and the pointer
size is tested from the compiler), but for arm architectures we
need to specify them explicitly.

Signed-off-by: Martin Storsjö <martin@martin.st>
---
Not sure if we want to do this, yet at least. These tests seem
to run into occasional spurious failures at times - not sure yet
why that is; it may be related to the msys2 shell and associated
msys2 binaries running emulated.