crt: Define fseeko64 symbol as alias to _fseeki64
mingw-w64 implementation of fseeko64() and _fseeki64() functions in files
mingw-w64-crt/stdio/_fseeki64.c and mingw-w64-crt/stdio/fseeko64.c are
exactly same.

Also both both fseeko64() and _fseeki64() functions have C fseek() ABI with
64-bit offset type.

So remove the mingw-w64 fseeko64 implementation and define fseeko64 symbol
as an alias to the _fseeki64 symbol. For msvcr80+ define it in def include
file crt-aliases.def.in and for other implementations define it in the
mingw-w64-crt/stdio/_fseeki64.c file.

This change allows to use native msvcr80+ / UCRT implementation of
_fseeki64() function for the POSIX fseeko64() instead of statically linking
the mingw-w64 implementation.

Signed-off-by: Martin Storsjö <martin@martin.st>