crt: Split testcases to a separate Makefile.am
This makes the global AM_CFLAGS in mingw-w64-crt not take effect
when compiling the testcases. This fixes building the testcases
for e.g. UCRT configurations, by using the installed toolchain's
default CRT configuration mode, rather than the specific
configuration used for building the mingw-w64-crt source files
(which are built with a fixed configuration).