crt: Add _setjmp and _setjmpex symbol aliases for UCRT builds
UCRT has _setjmp and _setjmpex function symbols in DLL libs prefixed by the
__intrinsic keyword. Add symbol aliases without this prefix which allows to
call _setjmp and _setjmpex without the prefix also for UCRT builds and so
allows to unify header files for msvcrt vs UCRT builds.

Aliases in crt-aliases.def.in are used for libucrtbase.a and libucrtbased.a
import libraries. Aliases in api-ms-win-crt-private-l1-1-0.def.in are used
for libucrt.a import library. And aliases in setjmp.S are used for
libucrtapp.a import library.

Signed-off-by: Martin Storsjö <martin@martin.st>