headers: Fix setjmpex.h support
When the msvc setjmpex.h file is included in source file then msvc changes
the behavior of C setjmp() and longjmp() calls to be SEH safe. Without
including the setjmpex.h file, these C calls are not SEH safe by default.

In following table is behavior of the msvc compiler, which symbols are
called in the emitted binary object code (not linked):

               without setjmpex.h         with setjmpex.h

    i386      _setjmp3  + longjmp      _setjmp3  + _longjmpex
    amd64     _setjmp   + longjmp      _setjmpex + longjmp
    arm32     _setjmp   + longjmp      _setjmpex + longjmp
    arm64     _setjmpex + longjmp      _setjmpex + longjmp

This table was produced from results on https://godbolt.org/z/Gabv1qT9c

When the compiled code is linked into final binary with UCRT library, the
_setjmp symbol is resolved to UCRT _intrinsic_setjmp symbol and the
_setjmpex symbol is resolved to UCRT _intrinsic_setjmpex. i386 _setjmp3
stays as without the _intrinsic prefix and the i386 _setjmp is not emitted
at all.

Fix the mingw-w64 setjmpex.h and setjmp.h include files to define setjmp()
and longjmp() macros which expands to functions emitted by msvc compiler to
match the msvc behavior.

Note that msvc setjmpex.h and setjmp.h header files are different than
mingw-w64 and is because it looks like that msvc compiler parses the source
code with setjmp and longjmp keywords differently and is doing some
transformations not specified in header files. Something which gcc is not
doing.

mingw-w64 setjmpex.h and setjmp.h files are defined based on the
observation of output from msvc compiler, not from the msvc header files.

Signed-off-by: Martin Storsjö <martin@martin.st>