crt: Make mingw-w64 implementation of wcstok compatible with MS implementation
MS implementation does not set *context to NULL, but instead sets it to the
pointer to nul char which is after the last char of the input string.

This change is needed for mingw-w64's wcstok_s which needs to distinguish
between invalid parameters (and call exception handler) and valid
parameters resulted from the last wcstok_s call of the last token.

Signed-off-by: Martin Storsjö <martin@martin.st>