crt: Enable -Wno-pragma-pack if supported
Check if the compiler supports -Wpragma-pack (Clang supports this),
and enable -Wno-pragma-pack in that case. This silences warnings
like these:

    ../stdio/mingw_pformat.c:86:10: warning: the current #pragma pack alignment value is modified in the included file [-Wpragma-pack]

The fact that an included file changed the pragma pack alignment
is something that generally can be suspicious and Clang thinks this
is worth warning about, but the Windows headers <pshpack1.h>
specifically do this thing.

To avoid having to try to silence the warning inline, disable
this warning entirely in configure.ac.

GCC silently accepts options like -Wno-<diag> - but if such an
option is included and GCC prints a warning, it prints an extra
note saying that it didn't recognize the -Wno-<diag> option and
it may have been intended for silencing this warning.

Therefore, for checking support for a -Wno-<diag> option, check
for the positive form even if we are going to apply the negative
form of it.

Signed-off-by: Martin Storsjö <martin@martin.st>