crt: Enable -Wno-array-bounds if supported
This is required for silencing warnings on modern GCC versions.

GCC tracks the bounds of objects; even though the C code casts
such references to pointers to other objects, GCC keeps track of
the size of the original object.

One such case is found in delayimp.c, where we reference
"extern IMAGE_DOS_HEADER __ImageBase". A pointer to the __ImageBase
symbol is used for locating the PIMAGE_NT_HEADERS struct (in
PinhFromImageBase), which is located after the IMAGE_DOS_HEADER.
GCC flags any accesses to this struct as being out of bounds
for the original IMAGE_DOS_HEADER object.

Unless there's a good way to silence this diagnostic e.g. for
specific casts, we'll have to disable the warning entirely.