crt: Enable -Wno-array-bounds if supported
This is required for silencing warnings on modern GCC versions.

GCC tracks the bounds of objects; even though the C code casts
such references to pointers to other objects, GCC keeps track of
the size of the original object.

One such case is found in delayimp.c, where we reference
"extern IMAGE_DOS_HEADER __ImageBase". A pointer to the __ImageBase
symbol is used for locating the PIMAGE_NT_HEADERS struct (in
PinhFromImageBase), which is located after the IMAGE_DOS_HEADER.
GCC flags any accesses to this struct as being out of bounds
for the original IMAGE_DOS_HEADER object.

    ../misc/delayimp.c: In function ‘__HrLoadAllImportsForDll’:
    ../misc/delayimp.c:258:10: warning: array subscript ‘struct _IMAGE_NT_HEADERS64[0]’ is partly outside array bounds of ‘IMAGE_DOS_HEADER[1]’ [-Warray-bounds=]
      258 |   if(pinh->OptionalHeader.DataDirectory[IMAGE_DIRECTORY_ENTRY_DELAY_IMPORT].Size) {
          |          ^~
    ../misc/delayimp.c:50:25: note: object ‘__ImageBase’ of size 64
       50 | extern IMAGE_DOS_HEADER __ImageBase;
          |                         ^~~~~~~~~~~
    ../misc/delayimp.c:260:45: warning: array subscript ‘struct _IMAGE_NT_HEADERS64[0]’ is partly outside array bounds of ‘IMAGE_DOS_HEADER[1]’ [-Warray-bounds=]
      260 |     pidd = (PCImgDelayDescr) PtrFromRVA(pinh->OptionalHeader.DataDirectory[IMAGE_DIRECTORY_ENTRY_DELAY_IMPORT].VirtualAddress);
          |                                             ^~
    ../misc/delayimp.c:52:53: note: in definition of macro ‘PtrFromRVA’
       52 | #define PtrFromRVA(RVA)   (((PBYTE)&__ImageBase) + (RVA))
          |                                                     ^~~
    ../misc/delayimp.c:50:25: note: object ‘__ImageBase’ of size 64
       50 | extern IMAGE_DOS_HEADER __ImageBase;
          |                         ^~~~~~~~~~~

Unless there's a good way to silence this diagnostic e.g. for
specific casts, we'll have to disable the warning entirely.

Signed-off-by: Martin Storsjö <martin@martin.st>