crt: test: t_tmpfile: Ignore chdir errors
Directory C:\Windows does not have to be accessible or does not have to
exist at all. For example when system is installed on D: letter or when old
pre-XP systems were installed in C:\WinNT directory.

Try to fallback to C:\ but in all cases ignore errors.
The whole test should work from any working directory.

Signed-off-by: Martin Storsjö <martin@martin.st>